fix:security,jwt 수정
- 예외처리
- Entity 제약조건 추가
-

+++ src/main/java/com/potential_radar/PR/user/controller/UserController.java
@@ -12,6 +12,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Map;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
@@ -21,7 +23,7 @@ public class UserController {
     private final TokenService tokenService;
 
     @PostMapping("/login")
-    public ResponseEntity<?> login(@RequestBody UserLoginRequest loginRequest) {
+    public ResponseEntity<Object> login(@RequestBody UserLoginRequest loginRequest) {
         try{
            LoginResponse tokens= userService.login(loginRequest);
 
@@ -29,27 +31,40 @@ public class UserController {
 
         }catch (IllegalArgumentException e){
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("로그인 처리 중 오류가 발생했습니다");
         }
     }
 
    @PostMapping("/signup")
-    public ResponseEntity<?> signup(@RequestBody UserSignupRequest request) {
+    public ResponseEntity<Object> signup(@RequestBody UserSignupRequest request) {
        try{
            User newUser = userService.register(request);
-           return ResponseEntity.ok().body("회원가입 성공");
-
+           return ResponseEntity.ok().body(Map.of("message", "회원가입 성공",
+                   "userId", newUser.getUserId()));
        }catch (IllegalArgumentException e){
            return ResponseEntity.badRequest().body(e.getMessage());
+       }catch (Exception e){
+           return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("회원가입 처리 중 오류가 발생했습니다");
        }
    }
 
 
    @PostMapping("/logout")
     public ResponseEntity<?> logout(@AuthenticationPrincipal User user) {
-        tokenService.deleteRefreshToken(user.getUserId());
-        return ResponseEntity.ok("로그아웃 성공");
+        try {
+            if (user == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증되지 않은 사용자입니다.");
+            }
+            tokenService.deleteRefreshToken(user.getUserId());
+            return ResponseEntity.ok("로그아웃 성공");
+
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("로그아웃 처리 중 오류가 발생했습니다");
+        }
    }
 
+   // 모든 인증이 필요한 요청에선 Authorization: Bearer {AccessToken} 헤더 사용
 
 
 }
